replaced u16 with size_t

+++ que.h
@@ -5,7 +5,7 @@
 
 // Circular Que (kMaxSize must be pow2)
 
-template <typename T, const uint16_t kMaxSize>
+template <typename T, const size_t kMaxSize>
 class Que {
 public:
 
@@ -22,11 +22,11 @@ public:
 		return write_pos_ != read_pos_;
 	}
 
-	inline uint16_t size() {
+	inline size_t size() {
 		return (write_pos_ - read_pos_) & (kMaxSize - 1);
 	}
 
-	inline uint16_t available_size() {
+	inline size_t available_size() {
 		return kMaxSize - size();
 	}
 
@@ -35,19 +35,19 @@ public:
 	}
 
 	inline void write(T value) {
-		uint16_t pos = write_pos_;
+		size_t pos = write_pos_;
 		data[pos] = value;
 		write_pos_ = (pos + 1) & (kMaxSize - 1);
 	}
 
 	inline T read() {
-		uint16_t pos = read_pos_;
+		size_t pos = read_pos_;
 		T value = data[pos];
 		read_pos_ = (pos + 1) & (kMaxSize - 1);
 		return value;
 	}
 
-	inline T peek(uint16_t pos = 0) {
+	inline T peek(size_t pos = 0) {
 		return data[(read_pos_ + pos) & (kMaxSize - 1)];
 	}
 
@@ -57,14 +57,14 @@ public:
 
 	inline void swallow() {
 		if (readable()) {
-			uint16_t pos = read_pos_;
+			size_t pos = read_pos_;
 			read_pos_ = (pos + 1) & (kMaxSize - 1);
 		}
 	}
 
 private:
-	volatile uint16_t write_pos_ = 0;
-	volatile uint16_t read_pos_ = 0;
+	volatile size_t write_pos_ = 0;
+	volatile size_t read_pos_ = 0;
 	T data[kMaxSize];
 };
